Guard deprecated APIs inside an `#ifndef TILEDB_REMOVE_DEPRECATIONS`.
